[FE] REFACTOR: admin info refactoring #985
